C.Generate

Documentation

class Annotate e where

Methods

annotate :: String -> e -> e

data Structure

Constructors

Structure

Fields

structureName :: Name

structureFields :: [(Name, Type)]

structureNeedsDiscriminator :: Bool

emit a macro that declares a discriminator when needed

structureHasDiscriminator :: Bool

the first field must appear first in the on memory layout, don't move it.

structureAligned :: Bool

this structure needs to be aligned to a pointer boundry, even if it woudn't be otherwise.

(=*) :: (ToExpression x, ToExpression y) => x -> y -> Statement

(&) :: (ToStatement a, ToStatement b) => a -> b -> Statement

(*#) :: (ToExpression a, ToExpression a1) => a -> a1 -> Expression

(+#) :: (ToExpression a, ToExpression a1) => a -> a1 -> Expression

class ToExpression a where

Methods

toExpression :: a -> Expression

class ToStatement a where

Methods

toStatement :: a -> Statement

eq :: (ToExpression x, ToExpression y) => x -> y -> Expression

basicStructure :: Structure

anonStructType :: [Type] -> Type

basicType :: String -> Type

basicGCType :: String -> Type

a basic type the garbage collector might want to follow, guarenteed to be the size of a pointer.

cast :: Type -> Expression -> Expression

noAssign :: Expression -> Expression

cif :: Expression -> Statement -> Statement -> Statement

constant :: Constant -> Expression

creturn :: Expression -> Statement

dereference :: Expression -> Expression

reference :: Expression -> Expression

class Draw d where

Methods

err :: String -> d

drawG :: Draw d => d -> Doc

emptyExpression :: Expression

enum :: Name -> Constant

data Expression

data Constant

expressionRaw :: String -> Expression

statementOOB :: Statement -> Statement

indexArray :: Expression -> Expression -> Expression

function :: Name -> Type -> [(Name, Type)] -> [FunctionOpts] -> Statement -> Function

data Function

functionCall :: Name -> [Expression] -> Expression

indirectFunctionCall :: Expression -> [Expression] -> Expression

data FunctionOpts

Constructors

Public

Attribute String

generateC

Arguments

:: [Function]
functions

-> [Structure]
extra structures

-> (Doc, Doc)
final result

goto :: Name -> Statement

subBlock :: Statement -> Statement

isEmptyExpression :: Expression -> Bool

label :: Name -> Statement

localVariable :: Type -> Name -> Expression

name :: String -> Name

data Name

newVar :: UniqueProducer m => Type -> m Expression

newDeclVar :: UniqueProducer m => Type -> m (Statement, Expression)

newAssignVar :: (Monad m, Draw d) => d -> Name -> Expression -> m Statement

newTmpVar :: (UniqueProducer m, Draw d) => d -> Expression -> m (Statement, Expression)

forLoop :: Expression -> Expression -> Expression -> Statement -> Statement

nullPtr :: Expression

number :: Integer -> Constant

floating :: Double -> Constant

operator :: (ToExpression a, ToExpression b) => String -> a -> b -> Expression

project :: Name -> Expression -> Expression

project' :: Name -> Expression -> Expression

projectAnon :: Int -> Expression -> Expression

ptrType :: Type -> Type

funPtrType :: Type -> [Type] -> Type

renderG :: Draw d => d -> String

sizeof :: Type -> Expression

data Statement

string :: String -> Expression

structAnon :: [(Expression, Type)] -> Expression

structType :: Name -> Type

switch' :: Expression -> [(Maybe Constant, Statement)] -> Statement

character :: Char -> Constant

toName :: [Char] -> Name

data Type

uoperator :: String -> Expression -> Expression

variable :: Name -> Expression

voidType :: Type

voidStarType :: Type