C.Prims

Documentation

data CallConv

Constructors

CCall

StdCall

CApi

Primitive

DotNet

data Safety

Constructors

Safe

Unsafe

newtype ExtType

Constructors

ExtType PackedString

newtype Requires

Constructors

Requires (Set (CallConv, PackedString))

data DotNetPrim

Constructors

DotNetField

DotNetCtor

DotNetMethod

primReqs :: Prim -> Requires

data Prim

Constructors

PrimPrim Atom

CConst

Fields

primRequires :: Requires

primConst :: !PackedString

Func

Fields

primRequires :: Requires

funcName :: !PackedString

primArgTypes :: [ExtType]

primRetType :: ExtType

primRetArgs :: [ExtType]

primSafety :: Safety

IFunc

Fields

primRequires :: Requires

primArgTypes :: [ExtType]

primRetType :: ExtType

AddrOf

Fields

primRequires :: Requires

primConst :: !PackedString

Peek

Fields

primArgTy :: Ty

Poke

Fields

primArgTy :: Ty

PrimTypeInfo

Fields

primArgTy :: Ty

primRetTy :: Ty

primTypeInfo :: !PrimTypeInfo

PrimString !PackedString

PrimDotNet

Fields

primStatic :: !Bool

primDotNet :: !DotNetPrim

primIOLike :: !Bool

primAssembly :: !PackedString

primDotNetName :: !PackedString

Op

Fields

primCOp :: Op Ty

primRetTy :: Ty

data PrimTypeInfo

Constructors

PrimSizeOf

PrimMaxBound

PrimMinBound

PrimAlignmentOf

PrimUMaxBound

primStaticTypeInfo :: Ty -> PrimTypeInfo -> Maybe Integer

primIsCheap :: Prim -> Bool

These primitives may safely be duplicated without affecting performance or correctness too adversly. either because they are cheap to begin with, or will be recombined in a later pass.

primIsConstant :: Prim -> Bool

whether a primitive represents a constant expression (assuming all its arguments are constant) TODO needs grin support

primEagerSafe :: Prim -> Bool

whether a primitive can be eagarly evaluated. TODO needs grin support

primPrim :: ToAtom a => a -> Prim

parseDotNetFFI :: Monad m => String -> m Prim

Generated by DrIFT : Look, but Don't Touch. *