Cmm.Op

Documentation

data BinOp

Constructors

Add

Sub

Mul

Mulx

UMulx

Div
round to -Infinity

Mod
mod rounding to -Infinity

Quot
round to 0

Rem
rem rounding to 0

UDiv
round to zero (unsigned)

UMod
unsigned mod

And

Or

Xor

Shl

Shr
shift right logical

Shra
shift right arithmetic

Rotl

Rotr

FAdd

FSub

FDiv

FMul

FPwr

FAtan2

Eq

NEq

Gt

Gte

Lt

Lte

UGt

UGte

ULt

ULte

FEq

FNEq

FGt

FGte

FLt

FLte

FOrdered

data UnOp

Constructors

Neg
2s compliment negation

Com
bitwise compliment floating

FAbs
floating absolute value

FNeg
floating point negation

Sin

Cos

Tan

Sinh

Cosh

Tanh

Asin

Acos

Atan

Log

Exp

Sqrt

Bswap
Switch the order of the bytes in a word

Ffs
Returns one plus the index of the least significant 1-bit of x, 0 if x is zero.

Clz
number of leading (from MSB) zeros, undefined if zero

Ctz
number of trailing (from LSB) zeros, undefined if zero.

Popcount
number of bits set to 1 in word

Parity
number of bits set to 1 mod 2

data ConvOp

Constructors

F2I
convert a floating point to an integral value via truncation

F2U
convert a floating point to an unsigned integral value via truncation, negative values become zero

U2F
convert an unsigned integral value to a floating point number

I2F
convert an integral value to a floating point number

F2F
convert a float from one precision to another, preserving value as much as possible

Lobits
extract the low order bits

Sx
sign extend a value (signed)

Zx
zero extend a value (unsigned)

I2I
perform a Lobits or a Sx depending on the sizes of the arguments

U2U
perform a Lobits or a Zx depending on the sizes of the arguments

B2B
a nop, useful for coercing hints (bits 2 bits)

data ValOp

Constructors

NaN

PInf

NInf

PZero

NZero

data ArchBits

Constructors

BitsMax

BitsPtr

BitsUnknown

data TyBits

Constructors

Bits !Int

BitsArch !ArchBits

BitsExt String

data TyHint

Constructors

HintSigned

HintUnsigned

HintFloat

HintCharacter

HintNone

data Ty

Constructors

TyBits !TyBits !TyHint

TyBool

TyComplex Ty

TyVector !Int Ty

preadTy :: ReadP Ty

readTy :: Monad m => String -> m Ty

stringToOpTy :: String -> Ty

bool :: Ty

bits_ptr :: Ty

bits_max :: Ty

bits8 :: Ty

bits16 :: Ty

bits32 :: Ty

bits64 :: Ty

class ToCmmTy a where

Methods

toCmmTy :: a -> Maybe Ty

cmmTyBits :: ToCmmTy a => a -> Maybe Int

cmmTyHint :: ToCmmTy a => a -> Maybe TyHint

data Op v

Constructors

BinOp BinOp v v

UnOp UnOp v

ValOp ValOp

ConvOp ConvOp v

binopType :: BinOp -> Ty -> Ty -> Ty

isCommutable :: BinOp -> Bool

commuteBinOp :: BinOp -> Maybe BinOp

isAssociative :: BinOp -> Bool

unopFloat :: Ty -> UnOp -> Maybe String

binopFunc :: Ty -> Ty -> BinOp -> Maybe String

binopInfix :: BinOp -> Maybe (String, Int)

class IsOperator o where

Methods

isCheap :: o -> Bool

isEagerSafe :: o -> Bool

Generated by DrIFT : Look, but Don't Touch. *