Doc.PPrint

Description

A Pretty printing class using multiparameter type classes for maximal generality with some useful instances.

the pprinted type comes as the last argument so newtype deriving can be used in more places.

Documentation

data Assoc

Constructors

AssocLeft

AssocRight

AssocNone

class DocLike d => PPrint d a where

Methods

pprint :: a -> d

pprintAssoc :: Assoc -> Int -> a -> d

pplist :: [a] -> d

pprintParen :: PPrint d a => a -> d

pprintPrec :: PPrint d a => Int -> a -> d

checkAssoc :: (Ord a1, DocLike a) => Assoc -> a1 -> Assoc -> a1 -> a -> a

checkAssocApp :: (Num a1, Ord a1, DocLike a) => Assoc -> a1 -> a -> a

pprintBinary :: (PPrint a a1, PPrint a a2) => Assoc -> Int -> Assoc -> Int -> a1 -> a -> a2 -> a