Doc.Pretty

Description

Modification od Daan Leijens pretty printer. main changes are use of Doc.DocLike framework and the addition of out-of-band data for html tags or ansi escape codes

Documentation

data Doc

putDoc :: Doc -> IO ()

hPutDoc :: Handle -> Doc -> IO ()

putDocM :: Monad m => (String -> m ()) -> Doc -> m ()

putDocMLn :: Monad m => (String -> m ()) -> Doc -> m ()

(</>) :: Doc -> Doc -> Doc

(<//>) :: Doc -> Doc -> Doc

(<$$>) :: Doc -> Doc -> Doc

sep :: [Doc] -> Doc

fillSep :: [Doc] -> Doc

hsep :: DocLike a => [a] -> a

vsep :: [Doc] -> Doc

cat :: [Doc] -> Doc

fillCat :: [Doc] -> Doc

hcat :: DocLike a => [a] -> a

vcat :: DocLike a => [a] -> a

align :: Doc -> Doc

hang :: Int -> Doc -> Doc

indent :: Int -> Doc -> Doc

fill :: Int -> Doc -> Doc

fillBreak :: Int -> Doc -> Doc

errorDoc :: Doc -> a

failDoc :: Monad m => Doc -> m a

softline :: Doc

softbreak :: Doc

line :: Doc

linebreak :: Doc

nest :: Int -> Doc -> Doc

group :: Doc -> Doc

column :: (Int -> Doc) -> Doc

nesting :: (Int -> Doc) -> Doc

width :: Doc -> (Int -> Doc) -> Doc

data SimpleDoc

Constructors

SEmpty

SChar Char SimpleDoc

SText !Int String SimpleDoc

SLine !Int SimpleDoc

renderPretty :: Float -> Int -> Doc -> SimpleDoc

renderCompact :: Doc -> SimpleDoc

displayS :: SimpleDoc -> ShowS

displayIO :: Handle -> SimpleDoc -> IO ()

displayM :: Monad m => (String -> m ()) -> SimpleDoc -> m ()

textProc :: (Char -> String) -> String -> Doc

oob :: String -> Doc

out of band data. This text will appear in the pretty printed output but won't count towards formatting, as far as pretty printing is concerned, it is of length 0.