E.E

Documentation

data Id

type IdMap = IntjectionMap Id

type IdSet = IntjectionSet Id

newIds :: IdSet -> [Id]

find some temporary ids that are not members of the set, useful for generating a small number of local unique names.

module E.Type

isWHNF :: E -> Bool

tIntegerzh :: E

tBits32zh :: E

tFunc :: E -> E -> E

tvrSilly :: TVr' E

tBoolzh :: E

lFalsezh :: Lit e E

lTruezh :: Lit e E

ePi :: TVr -> E -> E

eLam :: TVr -> E -> E

discardArgs :: Int -> E -> E

throw away first n EPi terms

tvrName :: Monad m => TVr -> m Name

tvrShowName :: TVr -> String

modAbsurd :: Module

modBox :: Module

nameConjured :: Module -> E -> Name

fromConjured :: Monad m => Module -> Name -> m E

isBottom :: E -> Bool

caseBodiesMapM :: Monad m => (E -> m E) -> E -> m E

caseBodiesMap :: (E -> E) -> E -> E

eToList :: Monad m => E -> m [E]

toString :: Monad m => E -> m [Char]

ltTuple :: [E] -> E

ltTuple' :: [E] -> E

p_unsafeCoerce :: Prim

p_dependingOn :: Prim

p_toTag :: Prim

p_fromTag :: Prim

fromUnboxedTuple :: Monad m => E -> m [E]

isUnboxedTuple :: E -> Bool

module E.FreeVars