E.FreeVars

Documentation

decomposeLet :: E -> ([Either (TVr, E) [(TVr, E)]], E)

pull apart an ELet and separate out recursive strongly connected components from an ELet.

decomposeDs :: [(TVr, E)] -> [Either (TVr, E) [(TVr, E)]]

separate out recursive strongly connected components from a declaration list

caseUpdate :: E -> E

freeIds :: E -> IdSet