E.FromHs

Documentation

convertDecls :: TiData -> IdMap Properties -> ClassHierarchy -> Map Name Type -> DataTable -> [HsDecl] -> IO [(Name, TVr, E)]

convertRules :: Module -> TiData -> ClassHierarchy -> Map Name Type -> DataTable -> [HsDecl] -> IO Rules

createInstanceRules :: Monad m => DataTable -> ClassHierarchy -> [(TVr, E)] -> m Rules

create a RULE for each instance attached to the class methods. These rules allow early specialization of monomorphic code, and are eventually used in E.TypeAnalysis.expandPlaceholder to fill out the generic class method bodies.

procAllSpecs :: Monad m => DataTable -> [Rule] -> [(TVr, E)] -> m ([(TVr, E)], Rules)

getMainFunction :: Monad m => DataTable -> Name -> Map Name (TVr, E) -> m (TVr, E)