E.Inline

Documentation

app :: MonadStats m => (E, [E]) -> m E

programMapRecGroups

Arguments

:: Monad m

=> IdMap (Maybe E)
initial map to apply

-> (Id -> Info -> m Info)
annotate based on Id map

-> (E -> Info -> m Info)
annotate letbound bindings

-> (E -> Info -> m Info)
annotate lambdabound bindings

-> ((Bool, [Comb]) -> m [Comb])
bool is true if group is recursive.

-> Program

-> m Program

Map recursive groups, allowing an initial map to be passed in and it will also propagate changes in the tvrInfo properly, and make sure nothing shadows one of the global names.

forceInline :: HasProperties a => a -> Bool

programDecomposedDs :: Program -> [Either (TVr, E) [(TVr, E)]]

programDecomposedCombs :: Program -> [(Bool, [Comb])]

programMapProgGroups

Arguments

:: Monad m

=> IdMap (Maybe E)
initial map to apply

-> (Program -> m Program)

-> Program

-> m Program

forceNoinline :: HasProperties a => a -> Bool

baseInlinability :: (Num a, HasProperties a1) => a1 -> E -> a

higher numbers mean we want to inline it more