E.SSimplify

Documentation

data Occurance

Constructors

Unused
unused means a var is not used at the term level, but might be at the type level

Once
Used at most once not inside a lambda or as an argument

OnceInLam
used once inside a lambda

ManyBranch
used once in several branches

Many
used many or an unknown number of times

LoopBreaker
chosen as a loopbreaker, never inline

cacheSimpOpts :: SimplifyOpts -> SimplifyOpts

simplifyE :: SimplifyOpts -> InE -> (Stat, OutE)

collectOccurance' :: E -> (E, IdMap UseInfo)

programPruneOccurance :: Program -> Program

programSSimplify :: SimplifyOpts -> Program -> Program

programSSimplifyPStat :: SimplifyOpts -> Program -> IO Program

data SimplifyOpts

Constructors

SimpOpts

Fields

so_noInlining :: !Bool

this inhibits all inlining inside functions which will always be inlined

so_finalPhase :: !Bool

no rules and don't inhibit inlining

so_postLift :: !Bool

don't inline anything that was lifted out

so_boundVars :: IdMap Comb

bound variables

so_forwardVars :: IdSet

variables that we know will exist, but might not yet.

so_boundVarsCache :: IdSet

so_cachedScope :: Env

emptySimplifyOpts :: SimplifyOpts