E.TypeAnalysis

Description

examine all uses of types in a program to determine which ones are actually needed in the method generation

Documentation

typeAnalyze :: Bool -> Program -> IO Program

type Typ = VMap () Name

expandPlaceholder :: Monad m => Comb -> m Comb