E.Values

Documentation

eTuple :: [E] -> E

eTuple' :: [E] -> E

unboxedTuple :: (CanType e, ~ * (TypeOf e) E) => [e] -> Lit e E

unboxedUnit :: E

unboxedTyUnit :: E

class ToE a where

Methods

toE :: a -> E

typeE :: a -> E

class ToEzh a where

Methods

toEzh :: a -> E

typeEzh :: a -> E

eCons :: E -> E -> E

eNil :: E -> E

emptyCase :: E

eCaseTup :: E -> [TVr] -> E -> E

eCaseTup' :: E -> [TVr] -> E -> E

eJustIO :: E -> E -> E

eCase :: E -> [Alt E] -> E -> E

eLet :: TVr -> E -> E -> E

This takes care of types right away, it simplifies various other things to do it this way.

eStrictLet :: TVr' E -> E -> E -> E

strict version of let, evaluates argument before assigning it.

substLet :: [(TVr, E)] -> E -> E

substLet' :: [(TVr, E)] -> E -> E

eLetRec :: [(TVr, E)] -> E -> E

prim_seq :: E -> E -> E

prim_unsafeCoerce :: E -> E -> E

from_unsafeCoerce :: Monad m => E -> m (E, E)

isState_ :: E -> Bool

unsafeCoerceOpt :: E -> (Int, E, E -> E)

isFullyConst :: E -> Bool

whether a value is a compile time constant

isAtomic :: E -> Bool

whether a value may be used as an argument to an application, literal, or primitive these may be duplicated with no code size or runtime penalty

isManifestAtomic :: E -> Bool

whether a type is obviously atomic. fast and lazy, doesn't recurse True -> definitely atomic False -> maybe atomic

isSmall :: E -> Bool

whether an expression is small enough that it can be duplicated without code size growing too much. (work may be repeated)

isCheap :: E -> Bool

whether an expression may be duplicated or pushed inside a lambda without duplicating too much work

isLifted :: E -> Bool

determine if term can contain _|_

whnfOrBot :: E -> Bool

isUnboxed :: E -> Bool

safeToDup :: E -> Bool

eToPat :: Monad m => E -> m (Lit TVr E)

patToE :: Monad m => Lit TVr E -> m E