Fixer.Fixer

Documentation

class Fixable a where

Fixable class, must satisfy the following rules

isBottom bottom == True x lub x == x x lub y == y lub x x lub bottom == x x minus bottom == x bottom minus x == bottom x minus y == z --> y lub z == x

Methods

bottom :: a

isBottom :: a -> Bool

lub :: a -> a -> a

minus :: a -> a -> a

lte :: a -> a -> Bool

showFixable :: a -> String

data Value a

data Rule

data Fixer

addRule :: MonadIO m => Rule -> m ()

add a rule to the current set

ioToRule :: IO () -> Rule

turn an IO action into a Rule

conditionalRule :: Fixable a => (a -> Bool) -> Value a -> Rule -> Rule

the function must satisfy the rule that if a >= b then f(a) implies f(b)

dynamicRule :: Fixable a => Value a -> (a -> Rule) -> Rule

findFixpoint :: MonadIO m => Maybe (String, Handle) -> Fixer -> m ()

find fixpoint, perhaps printing debugging information to specified handle. will not print anything if no calculation needed.

calcFixpoint :: MonadIO m => String -> Fixer -> m ()

isSuperSetOf :: Fixable a => Value a -> Value a -> Rule

modifiedSuperSetOf :: (Fixable a, Fixable b) => Value b -> Value a -> (a -> b) -> Rule

the function must satisfy the rule that if a >= b then f(a) >= f(b)

newFixer :: MonadIO m => m Fixer

ioValue :: IO (Value a) -> Value a

mainly for internal use

newValue :: (MonadIO m, Fixable a) => Fixer -> a -> m (Value a)

readValue :: (Fixable a, MonadIO m) => Value a -> m a

read result, calculating fixpoint if needed

readRawValue :: (Fixable a, MonadIO m) => Value a -> m a

value :: a -> Value a