Fixer.VMap

Documentation

data VMap p n

General data type for finding the fixpoint of a general tree-like structure.

data Proxy p

Constructors

Proxy p

DepthExceeded

vmapSingleton :: (Ord p, Ord n) => n -> VMap p n

vmapArgSingleton :: (Ord p, Ord n, Show p, Show n) => n -> Int -> VMap p n -> VMap p n

vmapArg :: (Ord p, Ord n, Show p, Show n) => n -> Int -> VMap p n -> VMap p n

vmapValue :: (Show p, Show n, Ord p, Ord n) => n -> [VMap p n] -> VMap p n

vmapMember :: Ord n => n -> VMap p n -> Bool

vmapProxyIndirect :: (Show p, Show n, Ord p, Ord n, Fixable (VMap p n)) => Int -> VMap p n -> VMap p n

vmapPlaceholder :: (Show p, Show n, Ord p, Ord n) => p -> VMap p n

vmapDropArgs :: Ord n => VMap p n -> VMap p n

vmapHeads :: Monad m => VMap p n -> m [n]