FlagDump

Documentation

data Flag

Flags

Constructors

AllTypes
show unified type table, after everything has been typechecked

Aspats
show as patterns

Atom
dump atom table on exit

Bindgroups
show bindgroups

BoxySteps
show step by step what the type inferencer is doing

C
don't delete C source file after compilation

Class
detailed information on each class

ClassSummary
summary of all classes

Core
show intermediate core code

CoreAfterlift
show final core before writing ho file

CoreBeforelift
show core before lambda lifting

CoreInitial
show core right after E.FromHs conversion

CoreMangled
de-typed core right before it is converted to grin

CoreMini
show details even when optimizing individual functions

CorePass
show each iteration of code while transforming

CoreSteps
show what happens in each pass

Datatable
show data table of constructors

DatatableBuiltin
show data table entries for some built in types

Dcons
data constructors

Decls
processed declarations

Defs
Show all defined names in a module

Derived
show generated derived instances

EAlias
show expanded aliases

EInfo
show info tags on all bound variables

ESize
print the size of E after each pass

EVerbose
print very verbose version of E code always

Exports
show which names are exported from each module

Grin
dump all grin to the screen

GrinDatalog
print out grin information in a format suitable for loading into a database

GrinFinal
final grin before conversion to C

GrinGraph
print dot file of final grin code to outputname_grin.dot

GrinInitial
grin right after conversion from core

GrinNormalized
grin right after first normalization

GrinPosteval
show grin code just before eval/apply inlining

GrinPreeval
show grin code just before eval/apply inlining

Imports
show in scope names for each module

Ini
all ini configuration options

Instance
show instances

Kind
show results of kind inference for each module

KindSteps
show steps of kind inference

OptimizationStats
show combined stats of optimization passes

Parsed
parsed code

Preprocessed
code after preprocessing/deliting

Program
impl expls, the whole shebang.

Progress
show basic progress indicators

Renamed
code after uniqueness renaming

Rules
show all user rules and catalysts

RulesSpec
show specialization rules

SccModules
show strongly connected modules in dependency order

Sigenv
initial signature environment

Srcsigs
processed signatures from source code

Stats
show extra information about stuff

Steps
show interpreter go

Tags
list of all tags and their types

The
'-d' flag. The following is a list of possible parameters you can pass to

Types
display unified type table containing all defined names

process :: Set Flag -> [[Char]] -> (Set Flag, [[Char]])

helpMsg :: [Char]

helpFlags :: [[Char]]