FrontEnd.KindInfer

Description

This module implements the Kind Inference algorithm, and the routines which use the product of kind inference to convert haskell source types into the simplified kind annotated types used by the rest of the FrontEnd.

Documentation

kiDecls :: KindEnv -> [HsDecl] -> IO KindEnv

data KindEnv

hsQualTypeToSigma :: Monad m => KindEnv -> HsQualType -> m Sigma

hsAsstToPred :: KindEnv -> HsAsst -> Pred

kiHsQualType :: KindEnv -> HsQualType -> KindEnv

kindOfClass :: Name -> KindEnv -> [Kind]

kindOf :: Name -> KindEnv -> Kind

restrictKindEnv :: (Name -> Bool) -> KindEnv -> KindEnv

hsTypeToType :: Monad m => KindEnv -> HsType -> m Type

getConstructorKinds :: KindEnv -> Map Name Kind