FrontEnd.ParseMonad

Description

Monads for the Haskell parser and lexer.

Parsing

data P a

Monad for parsing

data ParseResult a

The result of a parse.

Constructors

ParseOk a
The parse succeeded, yielding a value and a set of warnings.

ParseFailed SrcLoc String
The parse failed at the specified source location, with an error message.

atSrcLoc :: P a -> SrcLoc -> P a

data LexContext

Constructors

NoLayout

Layout Int

data ParseMode

Static parameters governing a parse. More to come later, e.g. literate mode, language extensions.

Constructors

ParseMode

Fields

parseFilename :: FilePath

original name of the file being parsed

parseOpt :: Opt

parseFFI :: Bool

parseUnboxedValues :: Bool

parseUnboxedTuples :: Bool

parseModeOptions :: Opt -> ParseMode

runParserWithMode :: ParseMode -> P a -> String -> ([Warning], ParseResult a)

runParser :: P a -> String -> ([Warning], ParseResult a)

getSrcLoc :: MonadSrcLoc m => m SrcLoc

setSrcLoc :: SrcLoc -> Lex a ()

pushCurrentContext :: P ()

popContext :: P ()

thenP :: P a -> (a -> P b) -> P b

returnP :: a -> P a

Lexing

data Lex r a

getInput :: Lex r String

discard :: Int -> Lex r ()

Discard some input characters (these must not include tabs or newlines).

lexNewline :: Lex a ()

Discard the next character, which must be a newline.

lexTab :: Lex a ()

Discard the next character, which must be a tab.

lexWhile :: (Char -> Bool) -> Lex a String

alternative :: Lex a v -> Lex a (Lex a v)

checkBOL :: Lex a Bool

setBOL :: Lex a ()

startToken :: Lex a ()

getOffside :: Lex a Ordering

pushContextL :: LexContext -> Lex a ()

popContextL :: String -> Lex a ()

lexParseMode :: Lex a ParseMode

pullCtxtFlag :: Lex a Bool

setFlagDo :: Lex a ()