FrontEnd.Rename

Documentation

renameModule :: MonadWarn m => Opt -> FieldMap -> [(Name, [Name])] -> HsModule -> m ((HsModule, [HsDecl]), Map Name Name)

unRename :: Name -> Name

unRename gets the original identifier name from the renamed version

collectDefsHsModule :: HsModule -> ([(Name, SrcLoc, [Name])], [(Name, Int)])

Collect all names defined in a module as well as their declaration points and any subnames they might have. In addition, collect the arities of any constructors.

data FieldMap

Constructors

FieldMap !(Map Name Int) !(Map Name [(Name, Int)])

class DeNameable a where

Methods

deName :: Module -> a -> a

renameStatement :: MonadWarn m => FieldMap -> [(Name, [Name])] -> Module -> HsStmt -> m HsStmt