FrontEnd.TypeSynonyms

Documentation

removeSynonymsFromType :: MonadWarn m => TypeSynonyms -> HsType -> m HsType

declsToTypeSynonyms :: MonadWarn m => TypeSynonyms -> [HsDecl] -> m TypeSynonyms

convert a set of type synonym declarations to a synonym map used for efficient synonym expansion

convert a set of type synonym declarations to a synonym map used for efficient synonym expansion, expanding out the body of synonyms along the way.

data TypeSynonyms

restrictTypeSynonyms :: (Name -> Bool) -> TypeSynonyms -> TypeSynonyms

showSynonyms :: DocLike d => (HsType -> d) -> TypeSynonyms -> d

showSynonym :: (DocLike d, Monad m) => (HsType -> d) -> Name -> TypeSynonyms -> m d