Grin.Optimize

Documentation

grinPush :: Stats -> Lam -> IO Lam

grinPush pushes the definitions of variables as far inward as they can go so peephole optimizations have a better chance of firing. when the order of definitons doesn't matter, it uses heuristics to decide which one to push to allow the most peephole optimizations.

grinSpeculate :: Grin -> IO Grin