Ho.Build

Documentation

module Ho.Type

dumpHoFile :: String -> IO ()

parseFiles

Arguments

:: Opt
Options to use when parsing files

-> [FilePath]
Targets we are building, used when dumping dependencies

-> [String]
Extra libraries to load

-> [Either Module FilePath]
Either a module or filename to find

-> (CollectedHo -> Ho -> IO CollectedHo)
Process initial ho loaded from file

-> (CollectedHo -> Ho -> TiData -> IO (CollectedHo, Ho))
Process set of mutually recursive modules to produce final Ho

-> IO (CompNode, CollectedHo)
Final accumulated ho

preprocess :: Opt -> FilePath -> ByteString -> IO ByteString

preprocessHs :: Opt -> FilePath -> ByteString -> IO ByteString

buildLibrary

Arguments

:: (CollectedHo -> Ho -> IO CollectedHo)

-> (CollectedHo -> Ho -> TiData -> IO (CollectedHo, Ho))
Process set of mutually recursive modules to produce final Ho

-> FilePath

-> IO ()