Ho.Collected

Documentation

data CollectedHo

the collected information that is passed around this is not stored in any file, but is what is collected from the ho files.

Constructors

CollectedHo

Fields

choExternalNames :: IdSet

this is a list of external names that are valid but that we may not know anything else about it is used to recognize invalid ids.

choCombinators :: IdMap Comb

these are the functions in Comb form.

choOrphanRules :: Rules

these are rules that may need to be retroactively applied to other modules

choHoMap :: Map ModuleGroup Ho

the hos

choLibDeps :: Map PackedString HoHash

libraries depended on

choHo :: Ho

these are caches of pre-computed values

choVarMap :: IdMap (Maybe E)

cache of variable substitution map

choDataTable :: CollectedHo -> DataTable

choClassHierarchy :: CollectedHo -> ClassHierarchy

choTypeSynonyms :: CollectedHo -> TypeSynonyms

choFixities :: CollectedHo -> FixityMap

choAssumps :: CollectedHo -> Map Name Type

choRules :: CollectedHo -> Rules

choEs :: CollectedHo -> [(TVr, E)]

updateChoHo :: CollectedHo -> CollectedHo