Ho.Type

Documentation

cff_magic :: ChunkType

cff_link :: ChunkType

cff_libr :: ChunkType

cff_jhdr :: ChunkType

cff_core :: ChunkType

cff_defs :: ChunkType

cff_lcor :: ChunkType

cff_ldef :: ChunkType

cff_idep :: ChunkType

cff_file :: ChunkType

type SourceHash = Hash

A SourceHash is the hash of a specific file, it is associated with a specific Module that said file implements.

type HoHash = Hash

HoHash is a unique identifier for a ho file or library.

type ModuleGroup = Module

while a Module is a single Module associated with a single haskell source file, a ModuleGroup identifies a group of mutually recursive modules. Generally it is chosen from among the Modules making up the group, but the specific choice has no other meaning. We could use the HoHash, but for readability reasons when debugging it makes more sense to choose an arbitrary Module.

data CollectedHo

the collected information that is passed around this is not stored in any file, but is what is collected from the ho files.

Constructors

CollectedHo

Fields

choExternalNames :: IdSet

this is a list of external names that are valid but that we may not know anything else about it is used to recognize invalid ids.

choCombinators :: IdMap Comb

these are the functions in Comb form.

choOrphanRules :: Rules

these are rules that may need to be retroactively applied to other modules

choHoMap :: Map ModuleGroup Ho

the hos

choLibDeps :: Map PackedString HoHash

libraries depended on

choHo :: Ho

these are caches of pre-computed values

choVarMap :: IdMap (Maybe E)

cache of variable substitution map

data HoHeader

The header contains basic information about the file, it should be enough to determine whether we can discard the file right away or consider it further.

Constructors

HoHeader

Fields

hohVersion :: Int

the version of the file format. it comes first so we don't try to read data that may be in a different format.

hohHash :: HoHash

my sha1 id

hohName :: Either ModuleGroup (PackedString, Version)

the human readable name, either the ModuleGroup or the library name and version.

hohLibDeps :: [(PackedString, HoHash)]

library dependencies

hohArchDeps :: [(PackedString, PackedString)]

arch dependencies, these say whether the file is specialized for a given arch.

data HoIDeps

These are the dependencies needed to check if a ho file is up to date. it only appears in ho files as hl files do not have source code to check against or depend on anything but other libraries.

Constructors

HoIDeps

Fields

hoIDeps :: Map SourceHash (Module, [(Module, SrcLoc)])

modules depended on indexed by a hash of the source.

hoDepends :: [(Module, SourceHash)]

Haskell Source files depended on

hoModDepends :: [HoHash]

Other objects depended on to be considered up to date.

hoModuleGroupNeeds :: [ModuleGroup]

library module groups needed

data HoLib

Constructors

HoLib

Fields

hoModuleMap :: Map Module ModuleGroup

arbitrary metainformation such as library author, web site, etc.

hoReexports :: Map Module Module

hoModuleDeps :: Map ModuleGroup [ModuleGroup]

hoMetaInfo :: [(PackedString, PackedString)]

data Library

Constructors

Library

Fields

libHoHeader :: HoHeader

libHoLib :: HoLib

libTcMap :: Map ModuleGroup HoTcInfo

libBuildMap :: Map ModuleGroup HoBuild

libExtraFiles :: [ExtraFile]

libFileName :: FilePath

data HoTcInfo

data only needed for type checking.

Constructors

HoTcInfo

Fields

hoExports :: Map Module [Name]

hoDefs :: Map Name (SrcLoc, [Name])

hoAssumps :: Map Name Type

used for typechecking

hoFixities :: FixityMap

hoKinds :: KindEnv

used for typechecking

hoTypeSynonyms :: TypeSynonyms

hoClassHierarchy :: ClassHierarchy

hoFieldMap :: FieldMap

data HoBuild

Constructors

HoBuild

Fields

hoDataTable :: DataTable

Filled in by E generation

hoEs :: [(TVr, E)]

hoRules :: Rules

data Ho

Constructors

Ho

Fields

hoModuleGroup :: ModuleGroup

hoTcInfo :: HoTcInfo

hoBuild :: HoBuild

data ExtraFile

Constructors

ExtraFile

Fields

extraFileName :: PackedString

extraFileData :: ByteString

Generated by DrIFT : Look, but Don't Touch. *

choCombinators_u :: (IdMap Comb -> IdMap Comb) -> CollectedHo -> CollectedHo

choExternalNames_u :: (IdSet -> IdSet) -> CollectedHo -> CollectedHo

choHo_u :: (Ho -> Ho) -> CollectedHo -> CollectedHo

choHoMap_u :: (Map ModuleGroup Ho -> Map ModuleGroup Ho) -> CollectedHo -> CollectedHo

choLibDeps_u :: (Map PackedString HoHash -> Map PackedString HoHash) -> CollectedHo -> CollectedHo

choOrphanRules_u :: (Rules -> Rules) -> CollectedHo -> CollectedHo

choVarMap_u :: (IdMap (Maybe E) -> IdMap (Maybe E)) -> CollectedHo -> CollectedHo

choCombinators_s :: IdMap Comb -> CollectedHo -> CollectedHo

choExternalNames_s :: IdSet -> CollectedHo -> CollectedHo

choHo_s :: Ho -> CollectedHo -> CollectedHo

choHoMap_s :: Map ModuleGroup Ho -> CollectedHo -> CollectedHo

choLibDeps_s :: Map PackedString HoHash -> CollectedHo -> CollectedHo

choOrphanRules_s :: Rules -> CollectedHo -> CollectedHo

choVarMap_s :: IdMap (Maybe E) -> CollectedHo -> CollectedHo

hoAssumps_u :: (Map Name Type -> Map Name Type) -> HoTcInfo -> HoTcInfo

hoClassHierarchy_u :: (ClassHierarchy -> ClassHierarchy) -> HoTcInfo -> HoTcInfo

hoDefs_u :: (Map Name (SrcLoc, [Name]) -> Map Name (SrcLoc, [Name])) -> HoTcInfo -> HoTcInfo

hoExports_u :: (Map Module [Name] -> Map Module [Name]) -> HoTcInfo -> HoTcInfo

hoFieldMap_u :: (FieldMap -> FieldMap) -> HoTcInfo -> HoTcInfo

hoFixities_u :: (FixityMap -> FixityMap) -> HoTcInfo -> HoTcInfo

hoKinds_u :: (KindEnv -> KindEnv) -> HoTcInfo -> HoTcInfo

hoTypeSynonyms_u :: (TypeSynonyms -> TypeSynonyms) -> HoTcInfo -> HoTcInfo

hoAssumps_s :: Map Name Type -> HoTcInfo -> HoTcInfo

hoClassHierarchy_s :: ClassHierarchy -> HoTcInfo -> HoTcInfo

hoDefs_s :: Map Name (SrcLoc, [Name]) -> HoTcInfo -> HoTcInfo

hoExports_s :: Map Module [Name] -> HoTcInfo -> HoTcInfo

hoFieldMap_s :: FieldMap -> HoTcInfo -> HoTcInfo

hoFixities_s :: FixityMap -> HoTcInfo -> HoTcInfo

hoKinds_s :: KindEnv -> HoTcInfo -> HoTcInfo

hoTypeSynonyms_s :: TypeSynonyms -> HoTcInfo -> HoTcInfo

hoDataTable_u :: (DataTable -> DataTable) -> HoBuild -> HoBuild

hoEs_u :: ([(TVr, E)] -> [(TVr, E)]) -> HoBuild -> HoBuild

hoRules_u :: (Rules -> Rules) -> HoBuild -> HoBuild

hoDataTable_s :: DataTable -> HoBuild -> HoBuild

hoEs_s :: [(TVr, E)] -> HoBuild -> HoBuild

hoRules_s :: Rules -> HoBuild -> HoBuild

hoBuild_u :: (HoBuild -> HoBuild) -> Ho -> Ho

hoModuleGroup_u :: (ModuleGroup -> ModuleGroup) -> Ho -> Ho

hoTcInfo_u :: (HoTcInfo -> HoTcInfo) -> Ho -> Ho

hoBuild_s :: HoBuild -> Ho -> Ho

hoModuleGroup_s :: ModuleGroup -> Ho -> Ho

hoTcInfo_s :: HoTcInfo -> Ho -> Ho