Info.Types

Description

some useful types to use in Info's that don't really fit anywhere else

Documentation

data Arity

how many arguments a function my be applied to before it performs work and whether it bottoms out after that many arguments

Constructors

Arity Int Bool

newtype Properties

list of properties of a function, such as specified by use pragmas or options

Constructors

Properties (EnumBitSet Property)

class HasProperties a where

Methods

modifyProperties :: (Properties -> Properties) -> a -> a

getProperties :: a -> Properties

putProperties :: Properties -> a -> a

setProperty :: Property -> a -> a

unsetProperty :: Property -> a -> a

getProperty :: Property -> a -> Bool

setProperties :: [Property] -> a -> a

fetchProperties :: Info -> Maybe Properties

module Info.Properties