Name.Id

Documentation

data Id

type IdMap = IntjectionMap Id

data IdNameT m a

Name monad transformer.

type IdSet = IntjectionSet Id

anonymous :: Int -> Id

va1 :: Id

some convinience anonymous ids

va2 :: Id

some convinience anonymous ids

va3 :: Id

some convinience anonymous ids

va4 :: Id

some convinience anonymous ids

va5 :: Id

some convinience anonymous ids

addBoundNamesIdMap :: Monad m => IntjectionMap Id a -> IdNameT m ()

addBoundNamesIdSet :: Monad m => IdSet -> IdNameT m ()

addNamesIdSet :: Monad m => IdSet -> IdNameT m ()

idMapToIdSet :: IdMap a -> IdSet

anonymousIds :: [Id]

sillyId :: Id

A occasionally useful random ethereal id

etherealIds :: [Id]

isEtherealId :: Id -> Bool

isInvalidId :: Id -> Bool

id isn't anonymous or atom-mapped

isEmptyId :: Id -> Bool

idSetToIdMap :: (Id -> a) -> IdSet -> IdMap a

mapMaybeIdMap :: (a -> Maybe b) -> IdMap a -> IdMap b

idSetFromList :: [Id] -> IdSet

idToInt :: Id -> Int

idSetFromDistinctAscList :: [Id] -> IdSet

idMapFromList :: [(Id, a)] -> IdMap a

idMapFromDistinctAscList :: [(Id, a)] -> IdMap a

idSetToList :: IdSet -> [Id]

idMapToList :: IdMap a -> [(Id, a)]

emptyId :: Id

newIds :: IdSet -> [Id]

find some temporary ids that are not members of the set, useful for generating a small number of local unique names.

newId

Arguments

:: Int
a seed value, useful for speeding up finding a unique id

-> (Id -> Bool)
whether an Id is acceptable

-> Id
your new Id

mixInt :: Int -> Int -> Int

mixInt3 :: Int -> Int -> Int -> Int

toId :: Name -> Id

fromId :: Monad m => Id -> m Name

candidateIds :: Int -> [Id]

runIdNameT :: Monad m => IdNameT m a -> m (a, IdSet)

Run the name monad transformer.