Util.NameMonad

Documentation

class Monad m => NameMonad n m | m -> n where

There are bound names and used names, the used names are always a superset of the bound names. used names will not be chosen for any new bindings, bound names should be renamed if encountered.

Methods

addNames :: [n] -> m ()

Add to list of used names

addBoundNames :: [n] -> m ()

Add to list of bound names

newName :: m n

Choose a new name, adding it to both bound and used sets.

newNameFrom :: [n] -> m n

choose the first available name from list

uniqueName :: n -> m n

choose a new name if n is bound, else return n adding n to the bound names list

class GenName n where

Generating names.

Methods

genNames :: Int -> [n]

Generate a list of candidate names given a seed

data NameMT n m a

Name monad transformer.

runNameMT :: Monad m => NameMT a1 m a -> m a

Run the name monad transformer.

runNameMT' :: Monad m => NameMT a1 m a -> m (a, Set a1)

freeNames :: (Ord n, GenName n) => Set n -> [n]

Generate an infinite list of names not present in the given set.

mixInt :: Int -> Int -> Int

mixInt3 :: Int -> Int -> Int -> Int

hashInt :: Int -> Int