Util.Once

Description

a simple type that only lets an IO action happen once, caching its result.

Documentation

data Once a

newOnce :: IO (Once a)

create a new Once object

runOnce :: Once a -> IO a -> IO a

execute the action at most once, always returning the same result

altOnce :: Once () -> IO b -> IO b -> IO b

run first argument once, after which perform the second

data OnceMap a b

run an IO action at most once for each distinct argument

newOnceMap :: Ord a => IO (OnceMap a b)

runOnceMap :: Ord a => OnceMap a b -> a -> IO b -> IO b

altOnceMap :: Ord a => OnceMap a () -> a -> IO b -> IO b -> IO b

onceMapToList :: OnceMap a b -> IO [(a, b)]

onceMapKeys :: OnceMap a b -> IO [a]

onceMapElems :: OnceMap a b -> IO [b]