Util.UniqueMonad

Documentation

data UniqT m a

Unique integer generator monad transformer.

type Uniq = UniqT Identity

Unique integer generator monad.

runUniq :: Int -> Uniq a -> (a, Int)

Run the bare version of the unique int generator.

runUniqT :: Monad m => UniqT m a -> Int -> m (a, Int)

Run the transformer version of the unique int generator.

execUniq1 :: Uniq a -> a

Execute the bare unique int generator starting with 1.

execUniq :: Int -> Uniq a -> a

Execute the bare unique int generator starting with the suplied number.

execUniqT :: Monad m => Int -> UniqT m a -> m a

Execute the transformer version of the unique int generator starting with the suplied number.